1.頁碼與每頁幾筆改成 int

+++ StockApi/Models/DataProviders/Stocks/HistoricalDailyQuoteParam.cs
@@ -5,8 +5,8 @@ namespace StockApi.Models.DataProviders.Stocks;
 internal struct HistoricalDailyQuoteParam(HistoricalDailyQuoteRequest req) : IKey
 {
     private string BaseKey { get; set; } = req.KeyWithPrefix();
-    public long PageIndex { get; set; } = req.RequestedPage;
-    public long PageSize { get; set; } = req.RecordsPerPage;
+    public int PageIndex { get; set; } = req.RequestedPage;
+    public int PageSize { get; set; } = req.RecordsPerPage;
     public DateOnly Date  { get; set; } = req.Date;
 
     public string KeyWithPrefix()
